Move update logic out of UseFormState again
