+++ frontend/composables/useFormState.ts
@@ -5,9 +5,6 @@ import type {
     FormWithValues,
 } from "~/composables/useProcessForm";
 import type { Ref } from "vue";
-import { graphql } from "~/generated/gql";
-import { useMutation } from "@vue/apollo-composable";
-import type { UpdateUserFormSubmission } from "~/generated/gql/graphql";
 
 interface FormState {
     formWithValues: FormWithValues | null;
@@ -15,27 +12,6 @@ interface FormState {
     formId: string | null;
 }
 
-const UPDATE_USER_FORM = graphql(`
-    mutation SaveFormSubmission(
-        $submissionId: ID
-        $formConfigId: ID
-        $responses: [ResponseInput!]!
-    ) {
-        updateFormSubmission(
-            userFormInput: {
-                submissionId: $submissionId
-                formConfigId: $formConfigId
-                responses: $responses
-            }
-        ) {
-            ok
-            errors {
-                field
-                messages
-            }
-        }
-    }
-`);
 
 export function useFormState(queriedFormRef: Ref<QueriedForm>) {
     const formId = queriedFormRef.value.formId;
@@ -71,34 +47,6 @@ export function useFormState(queriedFormRef: Ref<QueriedForm>) {
         { deep: true },
     );
 
-    const { mutate: mutateForm } = useMutation<UpdateUserFormSubmission>(
-        UPDATE_USER_FORM,
-        {
-            update: (cache) => {
-                cache.evict({ fieldName: "form" });
-                cache.gc();
-            },
-        },
-    );
-
-    function submitForm(): void {
-        const submissionId =
-            formState.value.formWithValues?.submissionId ?? null;
-        const formConfigId = formState.value.formId;
-        const formData = formState.value.formWithValues;
-        if (!formData) {
-            return;
-        }
-
-        const inputData = useFormDataToMutationInput(
-            formData,
-            submissionId,
-            formConfigId,
-        );
-        mutateForm(inputData).catch((error: unknown) => {
-            console.error("Error updating form:", error);
-        });
-    }
 
     return {
         formObject: computed({
@@ -108,6 +56,5 @@ export function useFormState(queriedFormRef: Ref<QueriedForm>) {
             },
         }),
         validationRules: computed(() => formState.value.validationRules),
-        submitForm,
     };
 }
